+++ bfkeyexp.py
@@ -1,4 +1,4 @@
-from __future__ import print_function
+
 import os
 from Crypto.Cipher import AES
 
@@ -15,15 +15,15 @@ def f_k_int_x(k, x):
         print("x+" + str(i) + ": Input to AES block " + str(i) + " encryption (128 bits):")
         xpi = "{0:032X}".format(x+i)
         print("0x" + xpi)
-        cArrayDef("[be]", "xp" + str(i), long(xpi, 16), 128/8, radix_8, False); print(os.linesep)
+        cArrayDef("[be]", "xp" + str(i), int(xpi, 16), 128/8, radix_8, False); print(os.linesep)
 
         print("AES_k(x+" + str(i) + "): Output of AES block " + str(i) + " encryption (128 bits):")
         aes_xpi = aes_obj.encrypt(xpi.decode('hex')).encode('hex')
         print("0x" + aes_xpi.upper())
-        cArrayDef("[be]", "aes_xp" + str(i), long(aes_xpi, 16), 128/8, radix_8, False); print(os.linesep)
+        cArrayDef("[be]", "aes_xp" + str(i), int(aes_xpi, 16), 128/8, radix_8, False); print(os.linesep)
 
         print("AES_k(x+" + str(i) + ") XOR (x+" + str(i) + "): block " + str(i) + " (128 bits):")
-        blki_int = long(xpi, 16) ^ long(aes_xpi, 16)
+        blki_int = int(xpi, 16) ^ int(aes_xpi, 16)
         blki = "{0:032X}".format(blki_int)
         print("0x" + blki)
         cArrayDef("[be]", "block_" + str(i), blki_int, 128/8, radix_8, False); print(os.linesep)
@@ -75,7 +75,7 @@ print("------------------------------------")
 
 print("ck: AES key (128 bits, randomly generated):")
 print("0x" + ck)
-cArrayDef("[be]", "ck", long(ck, 16), 128/8, radix_8, False); print(os.linesep)
+cArrayDef("[be]", "ck", int(ck, 16), 128/8, radix_8, False); print(os.linesep)
 
 print("a: Signing seed private key (256 bits):")
 print(Hex(a, radix_256))
@@ -94,10 +94,10 @@ print(Hex(x_cert, radix_128) + os.linesep)
 f_k_int_x_cert = f_k_int_x(ck, x_cert)
 print("f_k^{int}(x) = block1 || block2 || block3 (384 bits):")
 print("0x" + f_k_int_x_cert)
-cArrayDef("[be]", "f_k_int_x_cert", long(f_k_int_x_cert, 16), 384/8, radix_8, False); print(os.linesep)
+cArrayDef("[be]", "f_k_int_x_cert", int(f_k_int_x_cert, 16), 384/8, radix_8, False); print(os.linesep)
 
 print("f_k(x) = f_k^{int}(x) mod l, where l is the order of the group of points on the curve (256 bits):")
-f_k_x_cert = long(f_k_int_x_cert, 16) % genP256.ecc.n
+f_k_x_cert = int(f_k_int_x_cert, 16) % genP256.ecc.n
 print(Hex(f_k_x_cert, radix_256))
 cArrayDef("[le]", "f_k_x_cert", f_k_x_cert, 256/32, radix_32, True); print(os.linesep)
 
@@ -122,7 +122,7 @@ print("-----------------------------------")
 
 print("ek: AES key (128 bits, randomly generated):")
 print("0x" + ek)
-cArrayDef("[be]", "ek", long(ek, 16), 128/8, radix_8, False); print(os.linesep)
+cArrayDef("[be]", "ek", int(ek, 16), 128/8, radix_8, False); print(os.linesep)
 
 print("h: Signing seed private key (256 bits):")
 print(Hex(h, radix_256))
@@ -141,10 +141,10 @@ print(Hex(x_enc, radix_128) + os.linesep)
 f_k_int_x_enc = f_k_int_x(ek, x_enc)
 print("f_k^{int}(x) = block1 || block2 || block3 (384 bits):")
 print("0x" + f_k_int_x_enc)
-cArrayDef("[be]", "f_k_int_x_enc", long(f_k_int_x_enc, 16), 384/8, radix_8, False); print(os.linesep)
+cArrayDef("[be]", "f_k_int_x_enc", int(f_k_int_x_enc, 16), 384/8, radix_8, False); print(os.linesep)
 
 print("f_k(x) = f_k^{int}(x) mod l, where l is the order of the group of points on the curve (256 bits):")
-f_k_x_enc = long(f_k_int_x_enc, 16) % genP256.ecc.n
+f_k_x_enc = int(f_k_int_x_enc, 16) % genP256.ecc.n
 print(Hex(f_k_x_enc, radix_256))
 cArrayDef("[le]", "f_k_x_enc", f_k_x_enc, 256/32, radix_32, True); print(os.linesep)
